backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
supabase>=2.10.0
google-generativeai>=0.8.0
groq>=0.12.0
PyGithub>=2.5.0
pdfplumber>=0.11.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pillow>=10.0.0
